lifeplanner-agent/pyproject.toml
[project]
name = "lifeplanner-agent"
version = "0.1.0"
description = "家族向けライフプランナーエージェント(MF ME家計データ起点)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "charset-normalizer>=3.3.0",
    "sqlalchemy>=2.0.35",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "anyio[trio]>=4.4.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]
ignore = ["E501"]
